Add a means of listing installed software versions
We add a YAML-driven way to list installed software versions
at the end of the template createion.
The commands to list the versions of the desired software
will be indicated in each worker.*.yml file as the
'list_versions_by' array.

In general, this array will be constructed by listing
apt packages, and using appropriate shell commands.